Upload and download file by chunking

+++ PiscadaServer/src/apps/ApiServer/documents/documentapihandler.cpp
@@ -4,7 +4,7 @@
 #include "documents/documentservices.h"
 
 #include <QDebug>
-
+#include <QJsonObject>
 DocumentApiHandler::DocumentApiHandler(PiMqttClient *client) : TypedApiHandler(client)
 {
     m_handlers.insert("createFolder", &DocumentApiHandler::createFolder);
@@ -12,7 +12,10 @@ DocumentApiHandler::DocumentApiHandler(PiMqttClient *client) : TypedApiHandler(c
     m_handlers.insert("updateFolder", &DocumentApiHandler::updateFolder);
     m_handlers.insert("deleteFolder", &DocumentApiHandler::deleteFolder);
 
-    m_handlers.insert("createFile", &DocumentApiHandler::createFile);
+    m_handlers.insert("startFileUpload", &DocumentApiHandler::startFileUpload);
+    m_handlers.insert("uploadFileChunk", &DocumentApiHandler::uploadFileChunk);
+    m_handlers.insert("finishFileUpload", &DocumentApiHandler::finishFileUpload);
+
     m_handlers.insert("getFile", &DocumentApiHandler::getFile);
     m_handlers.insert("updateFile", &DocumentApiHandler::updateFile);
     m_handlers.insert("deleteFile", &DocumentApiHandler::deleteFile);
@@ -31,9 +34,8 @@ DocumentApiHandler::~DocumentApiHandler() = default;
 
 void DocumentApiHandler::createFolder(const PiMqttMessage &msg)
 {
-    DocumentCreateRequest request;
+    FolderCreateRequest request;
 
-    // Parse JSON from MQTT payload
     JsonError jsonError = JsonUtil::decode(msg.payload(), &request);
     if (jsonError.type() != JsonError::NoError)
     {
@@ -41,9 +43,9 @@ void DocumentApiHandler::createFolder(const PiMqttMessage &msg)
     }
 
     // Validate name uniqueness
-    if (DocumentDbRepo::doesNameExist(request.name, request.parentId, request.type))
+    if (DocumentDbRepo::doesNameExist(request.name, request.parentId, request.type, request.extension))
     {
-        return sendErrorResponse(msg, ApiError::conflictError("This name already exists in the parent folder"));
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "This name already exists in the parent folder"});
     }
 
     // Validate parent folder
@@ -56,11 +58,11 @@ void DocumentApiHandler::createFolder(const PiMqttMessage &msg)
     document.name = request.name;
     document.description = request.description;
     document.parentId = request.parentId;
-    document.type = request.type;
+    document.type = "folder";
     document.extension = request.extension;
     document.writeAccess = request.writeAccess;
     document.readAccess = request.readAccess;
-    document.id = QUuid::createUuid().toString(QUuid::WithoutBraces);
+    document.id = QUuid::createUuid().toString().remove("{").remove("}");
 
     // Insert the folder into the database
     ApiError error = DocumentDbRepo::create(document);
@@ -74,7 +76,6 @@ void DocumentApiHandler::createFolder(const PiMqttMessage &msg)
 
 void DocumentApiHandler::getFolder(const PiMqttMessage &msg)
 {
-    // Extract folder ID from the MQTT topic    
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -119,6 +120,12 @@ void DocumentApiHandler::updateFolder(const PiMqttMessage &msg)
         return sendErrorResponse(msg, error);
     }
 
+    // Check if the new name folder existed in parent folder
+    if ((document.name != request.name) && (DocumentDbRepo::doesNameExist(request.name, document.parentId, document.type, document.extension)))
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "This name already exists in the parent folder"});
+    }
+
     document.name = request.name;
     document.description = request.description;
     document.writeAccess = request.writeAccess;
@@ -136,7 +143,6 @@ void DocumentApiHandler::updateFolder(const PiMqttMessage &msg)
 
 void DocumentApiHandler::deleteFolder(const PiMqttMessage &msg)
 {
-    // Extract folder ID from the topic
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -152,6 +158,22 @@ void DocumentApiHandler::deleteFolder(const PiMqttMessage &msg)
         return sendErrorResponse(msg, error);
     }
 
+    // Delete all descendant files of the folder in system
+    QVector<DocumentDetail> descendantFiles;
+    error = DocumentDbRepo::getAllDescendantFiles(descendantFiles, id);
+    if (error.isError()) 
+    {
+        return sendErrorResponse(msg, error);
+    }
+    for (const DocumentDetail &document : descendantFiles) 
+    {
+        error = DocumentServices::deleteFileFromSystem(document.id, document.extension, DocumentServices::documentPath);
+        if (error.isError()) 
+        {
+            return sendErrorResponse(msg, error);  
+        }
+    }
+
     // Delete the folder from the database
     error = DocumentDbRepo::remove(id);
     if (error.isError())
@@ -162,10 +184,9 @@ void DocumentApiHandler::deleteFolder(const PiMqttMessage &msg)
     sendResponse(msg, document);
 }
 
+// Fetch all records from the documents table
 void DocumentApiHandler::listAllDocuments(const PiMqttMessage &msg)
 {
-    //Fetch all records from the documents table
-
     QVector<DocumentDetail> documents;
 
     ApiError error = DocumentDbRepo::list(documents, false);
@@ -179,7 +200,6 @@ void DocumentApiHandler::listAllDocuments(const PiMqttMessage &msg)
 
 void DocumentApiHandler::getChildrenOfFolder(const PiMqttMessage &msg)
 {
-    //Extract the folder ID from the topic
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -187,8 +207,7 @@ void DocumentApiHandler::getChildrenOfFolder(const PiMqttMessage &msg)
         return sendErrorResponse(msg, error);
     }
 
-    //Query the database for child documents of the given folder
-
+    // Query the database for child documents of the given folder
     QVector<DocumentDetail> documents;
     error = DocumentDbRepo::getChildrenOfFolder(id, documents);
 
@@ -200,65 +219,145 @@ void DocumentApiHandler::getChildrenOfFolder(const PiMqttMessage &msg)
     sendResponse(msg, documents);
 }
 
-void DocumentApiHandler::createFile(const PiMqttMessage &msg)
+void DocumentApiHandler::startFileUpload(const PiMqttMessage &msg)
 {
-    FileCreateRequest request;
-    // Parse JSON payload into FileCreateRequest
+    FileStartRequest request;
     JsonError jsonError = JsonUtil::decode(msg.payload(), &request);
     if (jsonError.type() != JsonError::NoError)
     {
         return sendErrorResponse(msg, ApiError::fromJsonError(jsonError));
     }
-    // Check if a file with the same name already exists in the parent folder
-    if (DocumentDbRepo::doesNameExist(request.name, request.parentId, request.type))
-    {
-        return sendErrorResponse(msg, ApiError::conflictError("This name already exists in the parent folder"));
-    }
 
-    // Validate that the parent folder exists
+    // Validate request parameters
+    if (DocumentDbRepo::doesNameExist(request.name, request.parentId, "file", request.extension))
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "Name already exists in the parent folder"});
+    }
     if (!DocumentDbRepo::isValidParent(request.parentId))
     {
         return sendErrorResponse(msg, ApiError::notFound("Parent folder", request.parentId));
     }
 
+    // Create id for the file upload session
+    QString id = QUuid::createUuid().toString().remove("{").remove("}");
+
+    QJsonObject result {
+        { "id", id },
+        { "name", request.name },
+        { "extension", request.extension },
+        { "totalChunks", request.totalChunks }
+    };
+    sendResponse(msg, result);
+}
+
+void DocumentApiHandler::uploadFileChunk(const PiMqttMessage &msg)
+{
+    FileChunkRequest request;
+    JsonError jsonError = JsonUtil::decode(msg.payload(), &request);
+    if (jsonError.type() != JsonError::NoError)
+    {
+        return sendErrorResponse(msg, ApiError::fromJsonError(jsonError));
+    }
+
+    // Validate request parameters
+    if (request.id.isEmpty())
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::InvalidRequest, "File's id is missing"});
+    }
+    if (request.chunkIndex < 0 || request.chunkIndex >= request.totalChunks)
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::InvalidRequest, "Invalid chunkIndex"});
+    }
+
+    // Validate mime type 
+    if (request.chunkIndex == 0)
+    {
+        ApiError mimeError = DocumentServices::validateMimeType(request.chunkBase64.toUtf8(), request.extension);
+        if (mimeError.isError())
+        {
+            return sendErrorResponse(msg, mimeError);
+        }
+    }
+
+    // Save the base64 chunk file to a temporary folder
+    ApiError saveError = DocumentServices::saveBase64ChunkToTemp(request.id, request.chunkIndex, request.chunkBase64);
+    if (saveError.isError())
+    {
+        return sendErrorResponse(msg, saveError);
+    }
+
+    QJsonObject result{
+        { "id", request.id },
+        { "name", request.name },
+        { "extension", request.extension },
+        { "totalChunks", request.totalChunks },
+        { "chunkIndex", request.chunkIndex },
+    };
+    sendResponse(msg, result);
+}
+
+// This function is called when all file chunks have been uploaded
+void DocumentApiHandler::finishFileUpload(const PiMqttMessage &msg)
+{
+    FileFinishRequest request;
+    JsonError jsonError = JsonUtil::decode(msg.payload(), &request);
+    if (jsonError.type() != JsonError::NoError)
+    {
+        return sendErrorResponse(msg, ApiError::fromJsonError(jsonError));
+    }
+
+    // Validate request parameters
+    if (request.id.isEmpty())
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::InvalidRequest, "File's id is missing"});
+    }
+    if (DocumentDbRepo::doesNameExist(request.name, request.parentId, "file", request.extension))
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "Name already exists in the parent folder"});
+    }
+    if (!DocumentDbRepo::isValidParent(request.parentId))
+    {
+        return sendErrorResponse(msg, ApiError::notFound("Parent folder", request.parentId));
+    }
+
+    // Merge all base64 chunks into a single file
+    ApiError mergeError = DocumentServices::mergeBase64ChunksToFile(request.id, request.extension, request.totalChunks);
+    if (mergeError.isError())
+    {
+        return sendErrorResponse(msg, mergeError);
+    }
+    DocumentServices::deleteTempFolder(request.id);
+
+    // Create a DocumentDetail object to save to the database
     DocumentDetail document;
+    document.id = request.id;
     document.name = request.name;
     document.description = request.description;
     document.parentId = request.parentId;
-    document.type = request.type;
+    document.type = "file";
     document.extension = request.extension;
-    document.writeAccess = request.writeAccess;
     document.readAccess = request.readAccess;
-    document.id = QUuid::createUuid().toString(QUuid::WithoutBraces);
-    
-    // Decode base64 content and validate MIME type
-    QByteArray decodedContent;
-    ApiError decodeError = DocumentServices::decodeAndValidateBase64File(request.fileContent.toUtf8(), document.extension, decodedContent);
-    if (decodeError.isError()) 
-    { 
-        return sendErrorResponse(msg, decodeError);
-    }
+    document.writeAccess = request.writeAccess;
 
-    // Save decoded file to the file system
-    ApiError fileError = DocumentServices::saveFileToSystem(document.id, document.extension, decodedContent);
-
-    if (fileError.isError())
-    {
-        return sendErrorResponse(msg, fileError);
-    }
-
-    // Store file to the database
     ApiError error = DocumentDbRepo::create(document);
     if (error.isError())
     {
+        DocumentServices::deleteFileFromSystem(document.id, document.extension, DocumentServices::documentPath);
         return sendErrorResponse(msg, error);
     }
-    sendResponse(msg, document);
+
+    QJsonObject result{
+        { "id", request.id },
+        { "name", request.name },
+        { "parentId", request.parentId },
+        { "extension", request.extension },
+        { "totalChunks", request.totalChunks },
+    };
+    return sendResponse(msg, result);
 }
 
 void DocumentApiHandler::getFile(const PiMqttMessage &msg)
 {
-    // Extract file ID from MQTT topic
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -274,31 +373,57 @@ void DocumentApiHandler::getFile(const PiMqttMessage &msg)
         return sendErrorResponse(msg, error);
     }
 
-    // Read file content from file system (base64 encoded)
-    QString fileContentBase64;
-    ApiError fileError = DocumentServices::readFileFromSystem(document.id, document.extension, fileContentBase64);
+    // Read file content from file system 
+    QByteArray fileData;
+    ApiError fileError = DocumentServices::readFileFromSystem(document.id, document.extension, &fileData);
     if (fileError.isError())
     {
         return sendErrorResponse(msg, fileError);
     }
 
-    FileResponse response;
-    response.id = document.id;
-    response.parentId = document.parentId;
-    response.name = document.name;
-    response.description = document.description;
-    response.type = document.type;
-    response.extension = document.extension;
-    response.writeAccess = document.writeAccess;
-    response.readAccess = document.readAccess;
-    response.fileContent = fileContentBase64;
+    // Encode to base64 and separate it
+    QByteArray base64Data = fileData.toBase64();
+    int chunkSize = 1024 * 1024; // 1MB
+    int totalChunks = base64Data.size() / chunkSize + (base64Data.size() % chunkSize != 0 ? 1 : 0);
 
-    sendResponse(msg, response);
+    for (int i = 0; i < totalChunks; ++i)
+    {
+        QByteArray chunk = base64Data.mid(i * chunkSize, chunkSize);
+
+        FileChunkResponse chunkResponse{
+            .id = document.id,
+            .parentId = document.parentId,
+            .name = document.name,
+            .description = document.description,
+            .type = document.type,
+            .extension = document.extension,
+            .writeAccess = document.writeAccess,
+            .readAccess = document.readAccess,
+            .chunkBase64 = QString::fromUtf8(chunk),
+            .chunkIndex = i,
+            .totalChunks = totalChunks
+        };
+
+        sendResponse(msg, chunkResponse);
+        qDebug() << "Sending chunk" << i << "/" << totalChunks << ":" << QString::fromUtf8(chunk);
+    }
+    QJsonObject result{
+        { "status", "completed" },
+        { "id", document.id },
+        { "name", document.name },
+        { "parentId", document.parentId },
+        { "extension", document.extension },
+        { "description", document.description },
+        { "type", document.type },
+        { "readAccess", document.readAccess },
+        { "writeAccess", document.writeAccess },
+        { "totalChunks", totalChunks },
+    };
+    return sendResponse(msg, result);
 }
 
 void DocumentApiHandler::updateFile(const PiMqttMessage &msg)
 {
-    // Extract file ID from MQTT topic
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -322,6 +447,12 @@ void DocumentApiHandler::updateFile(const PiMqttMessage &msg)
         return sendErrorResponse(msg, error);
     }
 
+    // Check if the file name already existed with same type in the parent folder
+    if ((document.name != request.name) && (DocumentDbRepo::doesNameExist(request.name, document.parentId, document.type, document.extension)))
+    {
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "This name already exists in the parent folder"});
+    }
+
     document.name = request.name;
     document.description = request.description;
     document.writeAccess = request.writeAccess;
@@ -339,7 +470,6 @@ void DocumentApiHandler::updateFile(const PiMqttMessage &msg)
 
 void DocumentApiHandler::deleteFile(const PiMqttMessage &msg)
 {
-    // Extract file ID from MQTT topic
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -355,7 +485,14 @@ void DocumentApiHandler::deleteFile(const PiMqttMessage &msg)
         return sendErrorResponse(msg, error);
     }
 
-    // Delete the file from the file system
+    // Delete  file from system
+    error = DocumentServices::deleteFileFromSystem(document.id, document.extension, DocumentServices::documentPath);
+    if (error.isError()) 
+    {
+        return sendErrorResponse(msg, error);
+    }
+
+    // Delete file from db
     error = DocumentDbRepo::remove(id);
     if (error.isError())
     {
@@ -368,7 +505,6 @@ void DocumentApiHandler::deleteFile(const PiMqttMessage &msg)
 // This function moves a document (a file or a folder) to the new destination folder
 void DocumentApiHandler::moveDocument(const PiMqttMessage &msg) {
 
-    // Extract the document ID from the topic
     DocumentMoveRequest request;
     JsonError jsonError = JsonUtil::decode(msg.payload(), &request);
 
@@ -393,9 +529,9 @@ void DocumentApiHandler::moveDocument(const PiMqttMessage &msg) {
     }
 
     // Check if the name already existed with same type in the destination folder
-    if (DocumentDbRepo::doesNameExist(document.name, request.parentId, document.type))
+    if (DocumentDbRepo::doesNameExist(document.name, request.parentId, document.type, document.extension))
     {
-        return sendErrorResponse(msg, ApiError::conflictError("This name already exists in the parent folder"));
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "This name already exists in the parent folder"});
     }
 
     // Update in the database
@@ -411,7 +547,6 @@ void DocumentApiHandler::moveDocument(const PiMqttMessage &msg) {
 
 void DocumentApiHandler::archiveFile(const PiMqttMessage &msg)
 {
-    // Extract the document ID from the topic
     QString id;
     ApiError error = getStringIdFromTopic(msg.topic(), id);
     if (error.isError())
@@ -439,7 +574,6 @@ void DocumentApiHandler::archiveFile(const PiMqttMessage &msg)
 
 void DocumentApiHandler::unarchiveFile(const PiMqttMessage &msg)
 {
-    // Parse the JSON payload 
     DocumentRestoreRequest request;
     JsonError jsonError = JsonUtil::decode(msg.payload(), &request);
     if (jsonError.type() != JsonError::NoError)
@@ -462,9 +596,9 @@ void DocumentApiHandler::unarchiveFile(const PiMqttMessage &msg)
     }
 
     // Check if a file with the same name exists in the destination folder
-    if (DocumentDbRepo::doesNameExist(document.name, request.parentId, document.type))
+    if (DocumentDbRepo::doesNameExist(document.name, request.parentId, document.type, document.extension))
     {
-        return sendErrorResponse(msg, ApiError::conflictError("This name already exists in the parent folder"));
+        return sendErrorResponse(msg, ApiError{ApiError::ConflictError, "This name already exists in the parent folder"});
     }
 
     // Restore the document in the database
@@ -477,10 +611,9 @@ void DocumentApiHandler::unarchiveFile(const PiMqttMessage &msg)
     sendResponse(msg, document);
 }
 
+// Retrieve a list of deleted (archived) files from the database
 void DocumentApiHandler::listArchivedFiles(const PiMqttMessage &msg)
 {
-    // Retrieve a list of deleted (archived) files from the database
-
     QVector<DocumentDetail> deletedFiles; 
 
     ApiError error = DocumentDbRepo::list(deletedFiles, true);
@@ -495,8 +628,8 @@ void DocumentApiHandler::listArchivedFiles(const PiMqttMessage &msg)
 
 void DocumentApiHandler::getFullTree(const PiMqttMessage &msg)
 {
-    // Retrieve a list of all documents from the database
 
+    // Retrieve a list of all documents from the database
     QVector<DocumentDetail> documentList;
 
     ApiError err = DocumentDbRepo::list(documentList, false);
